requirements.txt
pyserial>3.4
python >3
